+++ BlazorOptions/ViewModels/EChartOptions.cs
@@ -5,19 +5,19 @@ namespace BlazorOptions.ViewModels;
 
 public record EChartOptions(
     Guid PositionId,
-    double[] Prices,
+    decimal[] Prices,
     string[] Labels,
-    double? TemporaryPrice,
+    decimal? TemporaryPrice,
     IReadOnlyList<ChartCollectionSeries> Collections,
-    double YMin,
-    double YMax);
+    decimal YMin,
+    decimal YMax);
 
 public record ChartCollectionSeries(
     Guid CollectionId,
     string Name,
     string Color,
     bool IsVisible,
-    IReadOnlyList<double> ExpiryProfits,
-    IReadOnlyList<double> TheoreticalProfits,
-    double? TemporaryPnl,
-    double? TemporaryExpiryPnl);
+    IReadOnlyList<decimal> ExpiryProfits,
+    IReadOnlyList<decimal> TheoreticalProfits,
+    decimal? TemporaryPnl,
+    decimal? TemporaryExpiryPnl);
